Fix: Address final two co-pilot suggestions
1. Extract duplicate email staging logic to shared helper function
   - Created src/lib/email/refund-notification.ts with stageRefundNotificationEmail()
   - Removed duplicate implementation from webhook handler
   - Replaced 58-line inline implementation in refund processing with single function call
   - Ensures consistency and simplifies future updates

2. Rename misleading 'total' field to 'count' in logs API response
   - Changed from 'total' (implies database total) to 'count' (returned logs)
   - Added clarifying comment
